+++ public/css/main.css
@@ -6,9 +6,27 @@ NO SHADOW, WE DON'T NEED SHADOW ANYWAY.
 @import "buttons.css";
 @import "labels.css";
 
+html {
+    position: relative;
+    min-height: 100%;
+}
+
 body {
-    padding-top: 95px;
-    padding-bottom: 75px
+    margin-bottom: 20px;
+}
+
+.footer {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 0px; /* Set the fixed height of the footer here */
+    line-height: 60px; /* Vertically center the text there */
+    background-color: #f5f5f5;
+}
+
+.footer > .navbar {
+    margin-bottom: 0;
+    height: auto;
 }
 
 .container-fluid {
